MINOR: receiver: add the foreign and v6only flags
The new RX_O_FOREIGN and RX_O_V6ONLY flags are now derived from the
listener at the moment the ->bind() call is made. It's difficult to
do it earlier because the keyword parsers are called one at a time
without context so it's not possible to know whether v6only and v4v6
conflict for example, nor is it easy to know when all keywords were
parsed. But fixing the receiver before starting it is fine.