WIP: protocols/receivers:
It's still not possible to eliminate the listener for UDP just because
we don't store the I/O handler, the thread mask nor the context in the
receiver.

It doesn't make much sense to duplicate all these settings for each and
every receiver for a given bind_conf. There seems to be something wrong
with where the split is performed.

Maybe we should first split the bind_conf to keep only receiver options.
such as unix uid/gid/mode, interface, namespace, bind_thread etc. This
would then become the real bind_conf, and the rest would be the listen_conf.

The storage would look like this :

 proxy --+
         |
         V
     listen_conf  ---------> bind_conf ---> proto
        |||                      ^            |
        VVV                      |            V
     listeners ------------> receiver     addr family

Another option would be that the receivers are actually between the
bind_conf and the listeners. We'd have this :

  proxy ---->>> bind_conf ---> proto
   |||            ||| ^
   VVV            VVV |
  listener <--> receivers

The main issue is that we still need to support this syntax:

   bind ipv4@addr4:443,ipv6@addr6:443,quicv4@addr4:443,quicv6@addr6:443 ssl ...

While it's acceptable to change the current behavior and have a single
listener for all socket-based listeners (ipv4,ipv6,unix-stream etc),
it's not when we switch to a new protocol (quic) as the setup will not
be the same, and the receivers will be set up as UDP.

So it *looks* like we need to keep multiple listeners per bind_conf. It
looks like each listener will have a single receiver, even though this
relation is not even mandatory for the long term. What matters instead
is that the receiver knows how to find the listener and the settings
that are common to the bind line regarding receivers:
  - interface
  - netns
  - unix perms
  - thread_mask

 proxy --+
         |
         V
     bind_conf ------------> rx_settings
        |||                      ^
        VVV                      |
     listeners <------------ receiver ----> addr family
        |                                   /
        +-----------> stream protocol -----'---> ctrl