MINOR: stream: Pass an optional input buffer when a stream is created
It is now possible to set the buffer used by the channel request buffer when a
stream is created. It may be useful if input data are already received, instead
of waiting the first call to the mux rcv_buf() callback. This change is
mandatory to support H1 connection with no stream attached.

For now, the multiplexers don't pass any buffer. BUF_NULL is thus used to call
stream_create_from_cs().