MINOR: mux-h1: Add a flag to disable reads on the connection
H1C_F_DONT_READ must be set on the H1 conenction to don't read more data. This
flag replaces the H1C_F_IN_BUSY flag. Its name is a bit explicit. It is
automatically set on the backend side when the mux is created. It is safe to do
so because at this stage, the request has not yet been sent to the server. This
way, in h1_recv_allowed(), a test on this flag is enough to block the reads
instead of testing the H1 stream state on the backend side.