MINOR: mux-h1: Separate parsing and formatting errors at H1 stream level
Instead of using H1S flags to report an error on the request or the response,
independently it is a parsing or a formatting error, we now use a flag to report
parsing errors and another one to report formatting ones. This simplify the
message parsing. It is also easier to figure out what error happened when one of
this flag is set. The side may be deduced checking the H1C_F_IS_BACK flag.