MINOR: mux-h1: Add a rxbuf into the H1 stream
For now this buffer is not used. But it will be used to parse the headers, and
possibly the first block of data, when no stream is attached to the H1
connection. The aim is to use it to create the stream, thanks to recent changes
on the streams creation api.