MINOR: mux-h1: rework the h1_timeout_task() function
Mainly to make it easier to read. First of all, when a H1 connection is still
there, we check if the connection was stolen by another thread or not. If yes we
release the task and leave. Then we check if the task is expired or not. Only
expired tasks are considered. Finally, if a conn-stream is still attached to the
connection (H1C_F_CS_ATTACHED flag set), we return. Otherwise, the task and the
H1 connection are released.