MINOR: session: Add the idle duration field into the session
The idle duration between two streams is added to the session structure. It is
not necessarily pertinent on all protocols. In fact, it is only defined for H1
connections. It is the duration between two H1 transactions. But the
.get_cs_info() callback function on the multiplexers only exists because this
duration is missing at the session level. So it is a simplification opportunity
for a really low cost.

To reduce the cost, a hole in the session structure is filled by moving
.srv_list field at the end of the structure.