MINOR: mux-h1: Disable reads if an error was reported on the H1 stream
Don't try to read more data if a parsing or a formatting error was reported on
the H1 stream. There is no reason to continue to process the messages fo the
current connection in this case. If a parsing error occurs, it means the input
is invalid. If a formatting error occurs, it is an internal error and it is
probably safer to give up.