MINOR: mux-h1: Rework how shutdowns are handled
We now only rely on one flag to notify a shutdown. The shutdown is performed at
the connection level when there are no more pending outgoing data. So, it means it
is performed immediately if the output buffer is empty. Otherwise it is deferred
after the outgoing data are sent.

This simplify a bit the mux because there is now only one flag to check.