MINOR: mux-h1: Add a idle expiration date on the H1 connection
An idle expiration date is added on the H1 connection with the function to set
it depending on connection state. First, there is no idle timeout on backend
connections, For idle frontend connections, the http-request or keep-alive
timeout are used depending on which timeout is defined and if it is the first
request or not. For embryonic connections, the http-request is always used, if
defined. For attached or shutted down connections, no idle timeout is applied.

For now the idle expiration date is never set and the h1_set_idle_expiration
function remains unused.