MINOR: stick-tables: Add functions to update some values of a tracked counter
The cumulative numbers of http requests, http errors, bytes received and sent
and their respective rates for a tracked counters are now updated using specific
stream independent functions. These functions are used by the stream but the aim
is to allow the session to do so too. For now, there is no reason to perform
these updates from the session, except from the mux-h2 maybe. But, the mux-h1,
on the frontend side, will be able to return some errors to the client, before
the stream creation. In this case, it will be mandatory to update counters
tracked at the session level.