MINOR: session: Add functions to increase http values of tracked counters
cumulative numbers of http request and http errors of counters tracked at the
session level and their rates can now be updated at the session level thanks to
two new functions. These functions are not used for now, but it will be called
to keep tracked counters up-to-date if an error occurs before the stream
creation.