MINOR: logs: Get the multiplexer exist status when no stream is provided
When a log message is emitted from the session level, by a multiplexer, there is
no stream. Thus for HTTP session, there no status code and the termination flags
are not correctly set.

Thanks to previous patch, the HTTP status code is retrieved using the
MUX_EXIT_STATE ctl param. This is only done for HTTP frontends. If it is defined
( != 0), it is used to deduce the termination flags.