MINOR: mux-h1: Add functions to send HTTP errors from the mux
For now, this part is unsued. But this patch adds functions to handle errors on
idle and embryonic H1 connections and send corresponding HTTP error messages to
the client (400, 408 or 500). Thanks to previous patches, these functions take
care to update the right stats counters, but also the counters tracked by the
session.

A field to store the HTTP error code has been added in the H1C structure. It is
used for error retransmits, if any, and to get it in http logs. It is returned
when the MUX_EXIT_STATUS ctl parameter is requested.