MAJOR: mux-h1: Create the client stream as later as possible
This is the reason for all previous patches. The conn-stream and the associated
stream are created as later as possible. It only concerns the frontend
connections. But it means the request headers, and possibly the first data
block, are received and parsed before the conn-stream creation. To do so, an
embryonic H1 stream, with no conn-stream, is created. The result of this "early
parsing" is stored in its rx buffer, used to fill the request channel when the
stream is created. During this step, some HTTP errors may be returned by the
mux. It must also handle http-request/keep-alive timeouts. A significative
change is about H1 to H2 upgrade. It happens very early now, and no H1 stream
are created (and thus of course no conn-stream).

The most important part of this patch is located to the h1_process()
function. Because it must trigger the parsing when there is no H1
stream. h1_recv() function has also been simplified.