MEDIUM: http-ana: Don't process partial or empty request anymore
It is now impossible to start the HTTP request processing in the stream
analysers with a partial or empty request message. The mux-h2 was already
waiting of the request headers before creating the stream. Now the mux-h1 does
the same. All errors (aborts, timeout or invalid requests) waiting for the
request headers are now handled by the multiplexers. So there is no reason to
still handle them in the REQ_WAIT_HTTP (http_wait_for_request) analyser.

To ensure there is no ambiguity, a BUG_ON() was added to exit if a partial
request is received in this analyser.