BROKEN/MEDIUM: listeners: rework the unbind logic to make it idempotent
BROKEN: not setting ASSIGNED when leaving READY on !do_close causes
reg-tests/http-rules/acl_cli_spaces.vtc to fail! It looks like the
command is sent to the master CLI instead, which could indicate a
much deeper problem such as a shared FD or reuse of a structure
between the master and the worker!

do_unbind_listener() is not logical and is not even idempotent. It must
not touch the fd if already -1, which also means not touch the receiver.
In addition, when performing a partial stop on a socket (not closing),
we know the socket remains in the listening state yet it's marked as
LI_ASSIGNED, which is confusing as it doesn't translate its real state.

With this change, we make sure that FDs marked for close end up in
ASSIGNED state and that those which are really bound and on which a
listen() was made (i.e. not pause) remain in LISTEN state. This is what
is closest to reality.

Ideally this function should become a default proto->unbind() one but
it may still keep a bit too much state logic to become generalized to
other protocols (e.g. QUIC).