MEDIUM: http-ana: Deal with L7 retries in HTTP analysers
The code dealing with the copy of requests in the L7-buffer and the retransmits
during L7 retries has been moved in the HTTP analysers. The copy is now
performed in the REQ_HTTP_XFER_BODY analyser and the L7 retries is performed in
the RES_WAIT_HTTP analyser. This way, si_cs_recv() and si_cs_send() don't care
of it anymore. It is much more natural to deal with L7 retry in HTTP analysers.