MINOR: htx: Add a function to know if a block is the last one
The htx_is_last_blk() function may now be used to know if a block is the
last one in an HTX message. Note the purpose of this function is not to know
if a block is the end of the message or if it is the last one insrted in the
message. But only if a block is the last one and no other block is in the
message.