BUG/MEDIUM: h2/mux-h2: Properly handle tunnel establishment
Don't add ES flag on HEADERS frame if a EOM block is encountered for a TUNNEL
message (on a CONNECT request or a 2xx response to a CONNECT request). To do so,
the H2_SF_BODY_TUNNEL flag is set on a h2 stream if a tunnel should be
established.

In addition, an PROTOCOL_ERROR error is emitted if a DATA frame is received from
the client on a non-fully established tunnel.

This patch may probably be backported as far as 2.0 (or at least 2.2 ?), but
with caution.