MINOR: mux-h1: Reject messages if data are received unexpectedly
If, during the message parsing, a mux waiting for the other side receives data,
a parsing error is triggered. This may happens for tunneled data sent to early
for a CONNECT request waiting for the 2xx response. It is unexpected and
forbidden to receive raw data at this stage. The client must wait the server
response.